requirements.txt
albumentations==1.3.1
lightning==2.2.0.post0
numpy==1.26.4
omegaconf==2.3.0
opencv_python_headless==4.9.0.80
Pillow==10.3.0
pudb==2024.1
pytorch_msssim==1.0.0
PyYAML==6.0.1
Requests==2.31.0
scipy==1.13.0
skimage==0.0
termcolor==2.4.0
thop==0.1.1.post2209072238
torchmetrics==1.2.1
tqdm==4.66.2
wandb==0.16.3


